Migrated product lsit
# Conflicts:
#	src/Core/Grid/Query/Filter/SqlFilters.php

# Conflicts:
#	src/Adapter/Product/Grid/Data/Factory/ProductPriceGridDataFactoryDecorator.php
#	src/Adapter/Product/ImageThumbnailProvider.php
#	src/Adapter/Product/ProductDataProvider.php
#	src/Core/Grid/Data/Factory/ProductGridDataFactoryDecorator.php
#	src/Core/Grid/Definition/Factory/ProductGridDefinitionFactory.php
#	src/Core/Grid/Query/Filter/DoctrineFilterApplicator.php
#	src/Core/Grid/Query/ProductQueryBuilder.php
#	src/Core/Search/Filters/ProductFilters.php
#	src/PrestaShopBundle/Controller/Admin/Sell/Catalog/Product/ProductController.php
#	src/PrestaShopBundle/Resources/config/routing/admin/sell/catalog/product_v2/product.yml
#	src/PrestaShopBundle/Resources/config/services/core/grid/grid_definition_factory.yml
#	src/PrestaShopBundle/Resources/config/services/core/grid/grid_factory.yml

+++ src/Adapter/Product/ProductDataProvider.php
@@ -177,7 +177,7 @@ class ProductDataProvider
             $productId,
             true,
             null,
-            Configuration::get('PS_PRICE_DISPLAY_PRECISION'),
+            (int) Configuration::get('PS_PRICE_DISPLAY_PRECISION'),
             null,
             false,
             true,
